Validate that the user's input matches the given prompt.
We need the promptIndex in order to know which sentence difficulty to compare to so we need to give it a more global scope.

Note: It is possible to grab the sentence again using .val() but this is treating the DOM like a database and it is yet another jquery lookup (performance consideration). We have the data in pure javascript (not in the DOM) so we should use that whenever possible.